+++ src/App.tsx
@@ -23,27 +23,7 @@ import { Header } from './components/Header.tsx';
 import { PublicRiddlesPage } from './pages/PublicRiddlesPage.tsx';
 import { ReceivedRiddlesPage } from './pages/ReceivedRiddlesPage.tsx';
 import { MyRiddlesPage } from './pages/MyRiddlesPage.tsx';
-
-export const theme = createTheme({
-	palette: {
-		mode: 'dark',
-		background: {
-			default: '#001233'
-		},
-		primary: {
-			main: deepOrange[500] //TODO: replace
-		},
-		secondary: {
-			main: red[400]
-		},
-		text: {
-			//
-		}
-	},
-	typography: {
-		fontFamily: ['"Baloo 2"', 'Roboto'].join(', ')
-	}
-});
+import { theme } from './theme.ts';
 
 export const App = () => {
 	const x = 'AOJ ÉO!';
